requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
click==8.0.4
tqdm==4.65.0
easydict==1.9
fonttools==4.41.0
omegaconf==2.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.8.1
einops==0.6.1
tensorboard==2.14.0
tabulate==0.9.0
pycocotools
opencv-python-headless==4.8.0.74
pillow==9.3.0
scikit-image==0.21.0
albumentations==1.3.1
augly==1.0.0
scikit-learn==1.0.2
matplotlib==3.7.2
timm==0.9.2
kornia==0.6.12
func_timeout==4.3.5
thop
loguru

# faiss-gpu 1.7.3 python3.8
https://github.com/faiss-wheels/faiss-wheels/releases/download/v1.7.3/faiss_gpu-1.7.3-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
